requirements.txt
streamlit==1.5.1
requests==2.28.1
ppandas==1.3.4